backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
Flask-Bcrypt==1.0.1
PyMySQL==1.1.1
cryptography==42.0.8
PyJWT==2.8.0
bcrypt==4.1.3
cloudinary==1.40.0
sendgrid==6.11.0
gunicorn==22.0.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-flask==1.3.0
